Added bookings for customers and owners.

+++ routes/api/v1/hotel.php
@@ -5,11 +5,11 @@ use Illuminate\Support\Facades\Route;
 
 Route::prefix('hotel')->middleware(['auth:sanctum','verified-email'])->group(function () {
     Route::post('/', [HotelController::class, 'create'])->middleware('owner-role');
-    Route::post('/{id}/image', [HotelController::class, 'addImage'])->middleware('owner-role');
-    Route::delete('/{hotelId}/image/{imageId}', [HotelController::class, 'deleteImage'])->middleware('owner-role');
-    Route::get('/{id}', [HotelController::class, 'getById']);
-    Route::patch('/{id}', [HotelController::class, 'update'])->middleware('owner-role');
-    Route::delete('/{id}', [HotelController::class, 'delete'])->middleware('owner-role');
+    Route::post('/{hotelId}/image', [HotelController::class, 'addImage'])->middleware(['owner-role', 'check-hotel-ownership']);
+    Route::delete('/{hotelId}/image/{imageId}', [HotelController::class, 'deleteImage'])->middleware(['owner-role', 'check-hotel-ownership']);
+    Route::get('/{hotelId}', [HotelController::class, 'getById']);
+    Route::patch('/{hotelId}', [HotelController::class, 'update'])->middleware(['owner-role', 'check-hotel-ownership']);
+    Route::delete('/{hotelId}', [HotelController::class, 'delete'])->middleware(['owner-role', 'check-hotel-ownership']);
 });
 
 Route::get('/hotels', [HotelController::class, 'list'])->middleware('auth:sanctum','verified-email');
